requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
eventlet==0.33.3
Werkzeug==3.0.1
psycopg2-binary>=2.9.9
